position(update)

+++ src/gameAlgo/Position.java
@@ -1,15 +1,15 @@
 package gameAlgo;
 
 public class Position {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    static int between(Position x, Position y) {
-        return (int) Math.sqrt((x.x - y.x) * (x.x - y.x) + (x.y - y.y) * (x.y - y.y));
+    static double between (Position x, Position y) {
+        return Math.sqrt((x.x - y.x)*(x.x - y.x)  + (x.y - y.y)*(x.y - y.y));
     }
 }
